+++ client/src/stylesheets/styles.css
@@ -7,10 +7,40 @@ body {
   font-family: sans-serif;
 }
 
-.chart-container {
-  width: 50%;
-  height: 50%;
-  max-width: 600px;
-  max-height: 400px;
+header {
+  text-align: center;
+  padding: 20px;
+}
+
+.main-content {
+  display: flex;
+  justify-content: space-around;
+  align-items: flex-start;
+  padding: 20px;
+}
+
+.section-container {
+  width: 45%; /* Adjust the width as needed */
+  text-align: center;
+  padding: 20px;
+  border: 2px solid #000; /* Adding a black border outline */
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1); /* Adding a subtle shadow for better visibility */
+}
+
+.doughnut-chart-container {
+  width: 100%;
+  max-width: 500px;
   margin: 0 auto; /* Center the chart horizontally */
+}
+
+.bar-chart-container {
+  width: 100%;
+  height: 500px;
+  max-width: 500px;
+  margin: 0 auto; /* Center the chart horizontally */
+}
+
+button {
+  display: block;
+  margin: 10px auto; /* Center the button horizontally */
 }
